requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
aniso8601==9.0.1
anyio==3.7.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
azure-ai-formrecognizer==3.2.1
azure-cognitiveservices-speech==1.29.0
azure-common==1.1.28
azure-core==1.26.0
bentoml==1.0.23
blinker==1.6.2
boto3==1.28.27
botocore==1.31.27
build==0.10.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
ci-info==0.3.0
circus==0.18.0
click==8.1.3
click-option-group==0.5.6
cloudpickle==2.2.1
colorama==0.4.6
configobj==5.0.8
configparser==5.3.0
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.9
decorator==5.1.1
deeplake==3.6.19
deepmerge==1.1.0
Deprecated==1.2.14
dill==0.3.7
easyocr==1.4
entrypoints==0.4
etelemetry==0.3.0
exceptiongroup==1.1.1
filelock==3.12.2
Flask==2.0.2
flask-healthz==0.0.3
flask-restx==1.1.0
flask-selfdoc==1.4.0
flatbuffers==23.5.26
fonttools==4.40.0
frontend==0.0.3
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.6.0
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
httplib2==0.19.0
huggingface-hub==0.15.1
humbug==0.3.2
idna==2.10
imageio==2.31.1
importlib-metadata==6.0.1
importlib-resources==5.12.0
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
jmespath==1.0.1
jsonpickle==2.0.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
langchain==0.0.231
langchainplus-sdk==0.0.20
lazy_loader==0.2
libclang==16.0.0
looseversion==1.2.0
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.22.4
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opentelemetry-api==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-aiohttp-client==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pathlib==1.0.1
pathos==0.3.1
pathspec==0.11.1
Pillow==9.5.0
pip-requirements-parser==32.0.1
pip-tools==6.14.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.17.1
protobuf==4.23.2
prov==2.0.0
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.0
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
PyMuPDF==1.22.5
pynvml==11.5.0
pyodbc==4.0.39
pyparsing==2.4.7
pypdf==3.15.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxnat==1.5
PyYAML==5.4.1
pyzmq==25.1.0
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
s3transfer==0.6.2
safetensors==0.3.1
schema==0.7.5
scikit-image==0.21.0
scipy==1.11.0
seaborn==0.12.2
simple-di==0.1.5
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.28.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2023.4.12
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==2.0.2
torchvision==0.14.1
tornado==6.3.2
tqdm==4.65.0
traits==6.3.2
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.5
uvicorn==0.22.0
validators==0.20.0
waitress==2.1.2
watchdog==3.0.0
watchfiles==0.19.0
Werkzeug==2.3.6
whisper==1.1.10
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
